+++ app/src/main/java/org/breezyweather/common/basic/models/options/DarkModeLocation.kt
@@ -1,0 +1,52 @@
+/**
+ * This file is part of Breezy Weather.
+ *
+ * Breezy Weather is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation, version 3 of the License.
+ *
+ * Breezy Weather is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Breezy Weather. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.breezyweather.common.basic.models.options
+
+import android.content.Context
+import org.breezyweather.R
+import org.breezyweather.common.basic.models.options.basic.BaseEnum
+import org.breezyweather.common.basic.models.options.basic.UnitUtils
+
+enum class DarkModeLocation(
+    override val id: String,
+    val value: Boolean,
+) : BaseEnum {
+
+    DAY_NIGHT("day_night", true),
+    DARK("dark", false),
+    ;
+
+    companion object {
+
+        fun getInstance(
+            value: String,
+        ) = DarkModeLocation.entries.firstOrNull {
+            it.id == value
+        } ?: DARK
+
+        fun getInstance(
+            value: Boolean,
+        ) = DarkModeLocation.entries.firstOrNull {
+            it.value == value
+        } ?: DARK
+    }
+
+    override val valueArrayId = R.array.dark_mode_location_values
+    override val nameArrayId = R.array.dark_modes_location
+
+    override fun getName(context: Context) = UnitUtils.getName(context, this)
+}